fix: invalid text 'null' in request when no args given
In these requests, if the optional args are None, instead of passing a empty request body, a text 'null' will be added, which results in 400 response saying
```
{
    "error": {
        "message": "JSON failed to parse.",
        "status_code": 400
    }
}
```

This fix checks for the optional args, and only write to the request body if they have values.

It is mainly for `finish_file_data_upload` as it allows an emty body, but the other two will still fail anyway.